Changes regarding look up of nodes by using ansible and network names in ansible hosts

+++ system-tests/src/test/groovy/utils/TestnetNodes.groovy
@@ -1,5 +1,6 @@
 package utils
 
+import com.google.common.collect.ImmutableSet
 import com.radixdlt.test.SystemCounters
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
@@ -8,12 +9,12 @@ class TestnetNodes {
 
     private static final TestnetNodes instance = new TestnetNodes();
     private static final Logger logger = LogManager.getLogger(TestnetNodes.class);
-
+    private static final String scheme = "https"
     private TestnetNodes() {
     }
 
 
-    private List nodes
+    private List<String> nodes
     String keyVolume = "key-volume"
     private String sshDestinationLocDir = "/ansible/ssh"
     private String sshDestinationFileName = "testnet"
@@ -24,11 +25,7 @@ class TestnetNodes {
     }
 
     String getNodesURls() {
-        if (nodes == null) {
-            fetchNodes()
-        } else {
-            logger.info("Nodes information already present ${nodes}")
-        }
+        fetchNodes()
         return nodes.inject("", {
             str, node ->
                 if (this.nodes.first() == node)
@@ -37,19 +34,31 @@ class TestnetNodes {
         })
     }
 
-    private fetchNodes() {
+    ImmutableSet<String> nodeURLList() {
+        fetchNodes()
 
-        String clusterName = Optional.ofNullable(System.getenv("TESTNET_NAME")).orElse("testnet_2")
-        logger.info("Node information not avaliable. Fetching using ansible")
-        String sshKeylocation = Optional.ofNullable(System.getenv("SSH_IDENTITY")).orElse(System.getenv("HOME") + "/.ssh/id_rsa")
-        CmdHelper.runCommand("docker container create --name dummy -v ${keyVolume}:${sshDestinationLocDir} curlimages/curl:7.70.0")
-        CmdHelper.runCommand("docker cp ${sshKeylocation} dummy:${sshDestinationLocDir}/${sshDestinationFileName}")
-        CmdHelper.runCommand("docker rm -f dummy")
-        def output, error
-        (output, error) = CmdHelper.runCommand("docker run --rm  -v ${keyVolume}:/ansible/ssh --name node-ansible ${ansibleImage}  "
-                + "check.yml --limit ${clusterName} --list-hosts")
-        nodes = output.findAll({
-            !(it.contains("play") || it.contains("pattern") || it.contains("hosts") || it == "")
-        }).collect({ it.trim() })
+        return ImmutableSet.copyOf(this.nodes.collect {
+            "${scheme}://${it}" as String
+        });
+    }
+
+    private fetchNodes() {
+        if (nodes == null) {
+            String clusterName = Optional.ofNullable(System.getenv("TESTNET_NAME")).orElse("testnet_2")
+            logger.info("Node information not avaliable. Fetching using ansible")
+            String sshKeylocation = Optional.ofNullable(System.getenv("SSH_IDENTITY")).orElse(System.getenv("HOME") + "/.ssh/id_rsa")
+            CmdHelper.runCommand("docker container create --name dummy -v ${keyVolume}:${sshDestinationLocDir} curlimages/curl:7.70.0")
+            CmdHelper.runCommand("docker cp ${sshKeylocation} dummy:${sshDestinationLocDir}/${sshDestinationFileName}")
+            CmdHelper.runCommand("docker rm -f dummy")
+            def output, error
+            (output, error) = CmdHelper.runCommand("docker run --rm  -v ${keyVolume}:/ansible/ssh --name node-ansible ${ansibleImage}  "
+                    + "check.yml --limit ${clusterName} --list-hosts")
+            nodes = output.findAll({
+                !(it.contains("play") || it.contains("pattern") || it.contains("hosts") || it == "")
+            }).collect({ it.trim() })
+
+        } else {
+            logger.info("Nodes information already present ${nodes}")
+        }
     }
 }
